Integrate app menu example into examples/menu.py
- Added open_preferences() and check_for_updates() functions
- Created macos_app_menu using Menu('__app__', [...])
- Added inline comments explaining macOS-only behavior
- Removed standalone app_menu_test.py file
